[ADD] Added the AWS_DELETE_ON_DBDROP config variable

+++ attachment_s3/controllers/main.py
@@ -1,5 +1,5 @@
 import logging
-import boto3
+import os
 
 from odoo.addons.web.controllers.main import Database
 from odoo import http
@@ -12,6 +12,9 @@ _logger = logging.getLogger(__name__)
 class Database(Database):
     @http.route()
     def drop(self, master_pwd, name):
+        delete_on_drop = os.environ.get("AWS_DELETE_ON_DBDROP")
+        if not delete_on_drop:
+            return super(Database, self).drop(master_pwd, name)
         try:
             bucket = request.env["ir.attachment"]._get_s3_bucket()
             bucket.objects.all().delete()
